cloud_run/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.0
google-auth==2.23.0
requests==2.31.0
